eval: add explicit casts for unsigned to float conversions
Fixes -Wsign-conversion warnings.

+++ libnopegl/src/eval.c
@@ -111,14 +111,14 @@ static float f_smoothstep(float a, float b, float x)
 }
 
 /* See https://bitbashing.io/comparing-floats.html */
-static float f_close_p(float a, float b, float p) { return fabsf(a - b) <= p * f_max(fabsf(a), fabsf(b)); }
+static float f_close_p(float a, float b, float p) { return (float)(fabsf(a - b) <= p * f_max(fabsf(a), fabsf(b))); }
 static float f_close(float a, float b) { return f_close_p(a, b, 1e-6f); }
 
-static float f_eq(float a, float b) { return a == b; }
-static float f_gt(float a, float b) { return a > b; }
-static float f_gte(float a, float b) { return a >= b; }
-static float f_lt(float a, float b) { return a < b; }
-static float f_lte(float a, float b) { return a <= b; }
+static float f_eq(float a, float b) { return (float)(a == b); }
+static float f_gt(float a, float b) { return (float)(a > b); }
+static float f_gte(float a, float b) { return (float)(a >= b); }
+static float f_lt(float a, float b) { return (float)(a < b); }
+static float f_lte(float a, float b) { return (float)(a <= b); }
 
 static float f_isfinite(float x) { return (float)isfinite(x); }
 static float f_isinf(float x) { return (float)isinf(x); }
